# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/candidatures": {
            "get": {
                "summary": "Liste des candidatures",
                "description": "Récupère la liste de toutes les candidatures.",
                "operationId": "61862138c083897914493911367a271e",
                "responses": {
                    "200": {
                        "description": "Liste des candidatures récupérée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "candidate_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "annonce_id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "pending"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-03-24T15:11:09.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-03-24T15:11:09.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Créer une candidature",
                "description": "Permet à un candidat de postuler à une annonce.",
                "operationId": "faf5f149632d99271a3677ed5ce79bae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "candidate_id",
                                    "annonce_id",
                                    "status"
                                ],
                                "properties": {
                                    "candidate_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "annonce_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidature créée avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "candidate_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "annonce_id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-24T15:11:09.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-03-24T15:11:09.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation errors"
                                        },
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": [
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "property": "candidate_id"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/candidatures/status/{status}": {
            "get": {
                "summary": "Récupérer les candidatures par statut",
                "description": "Récupère toutes les candidatures ayant un statut spécifique.",
                "operationId": "5d31beff0683b9958f7deea4219093ff",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Statut des candidatures à récupérer (ex: pending, accepted, rejected)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidatures récupérées avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidature by status"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "candidate_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "annonce_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-24T15:11:09.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-03-24T15:11:09.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}